+++ frontend/src/Navigation.svelte
@@ -17,23 +17,23 @@
   } from '@sveltestrap/sveltestrap'
   import { profile } from './api/profile.svelte'
   import { _ } from './lib/Translate.svelte'
-  import Configuration from './configuration/Index.svelte'
-  import SiteTunnel from './siteTunnel/Index.svelte'
-  import StarrApps from './starrApps/Index.svelte'
-  import DownloadApps from './downloadApps/Index.svelte'
-  import MediaApps from './mediaApps/Index.svelte'
-  import SnapshotApps from './snapshotApps/Index.svelte'
-  import FileWatcher from './fileWatcher/Index.svelte'
-  import Endpoints from './endpoints/Index.svelte'
-  import Commands from './commands/Index.svelte'
-  import ServiceChecks from './serviceChecks/Index.svelte'
-  import Triggers from './triggers/Index.svelte'
-  import Integrations from './integrations/Index.svelte'
-  import Monitoring from './monitoring/Index.svelte'
-  import Metrics from './metrics/Index.svelte'
-  import LogFiles from './logFiles/Index.svelte'
-  import System from './system/Index.svelte'
-  import Profile from './profile/Index.svelte'
+  import Configuration from './pages/configuration/Index.svelte'
+  import SiteTunnel from './pages/siteTunnel/Index.svelte'
+  import StarrApps from './pages/starrApps/Index.svelte'
+  import DownloadApps from './pages/downloadApps/Index.svelte'
+  import MediaApps from './pages/mediaApps/Index.svelte'
+  import SnapshotApps from './pages/snapshotApps/Index.svelte'
+  import FileWatcher from './pages/fileWatcher/Index.svelte'
+  import Endpoints from './pages/endpoints/Index.svelte'
+  import Commands from './pages/commands/Index.svelte'
+  import ServiceChecks from './pages/serviceChecks/Index.svelte'
+  import Triggers from './pages/triggers/Index.svelte'
+  import Integrations from './pages/integrations/Index.svelte'
+  import Monitoring from './pages/monitoring/Index.svelte'
+  import Metrics from './pages/metrics/Index.svelte'
+  import LogFiles from './pages/logFiles/Index.svelte'
+  import System from './pages/system/Index.svelte'
+  import Profile from './pages/profile/Index.svelte'
   import Landing from './Landing.svelte'
   import { ltrim } from './lib/util'
   import { darkMode, toggleDarkMode } from './lib/darkmode.svelte'
