+++ frontend/src/includes/Footer.svelte
