--- frontend/src/mediaApps/Index.svelte
@@ -1,17 +1,0 @@
-<script lang="ts">
-  import { Card, CardHeader, CardBody, CardFooter } from '@sveltestrap/sveltestrap'
-  import { darkMode } from '../lib/darkmode.svelte'
-  $: theme = $darkMode ? 'dark' : 'light'
-</script>
-
-<Card class="mb-2" {theme}>
-  <CardHeader>
-    <h2>Media Apps</h2>
-  </CardHeader>
-  <CardBody>
-    <p>Configure and manage media server applications like Plex, Jellyfin, and Emby.</p>
-  </CardBody>
-  <CardFooter>
-    <small class="text-muted">Manage your media server integrations</small>
-  </CardFooter>
-</Card>
